+++ mkdocs_tags/plugin.py
@@ -11,13 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""A tags plugin for MkDocs
+"""An MkDocs plugin for tags support
 
-This plugin reads tags info from the metadata section from each
-page, generate a list of tags on each page, and generate a tags page
-with a list of all tags and the pages under each of them.
+This plugin reads tag info from the metadata section of each page,
+generate a list of tags on each page, and generate a tag page with a
+list of all tags and the pages under each of them.
 
-Not intended for direct usage.
+Not intended for direct import.
 """
 
 import copy
@@ -33,11 +33,11 @@ from typing import Dict
 _TAGS_META_ENTRY = 'tags'
 _ON_PAGE_TMPLT_CFG_ENTRY = 'on_page_tmplt'
 _ON_PAGE_TMPLT_PATH_CFG_ENTRY = 'on_page_tmplt_path'
-_TAGS_PAGE_TMPLT_CFG_ENTRY = 'tags_page_tmplt'
-_TAGS_PAGE_TMPLT_PATH_CFG_ENTRY = 'tags_page_tmplt_path'
-_TAGS_PAGE_MD_PATH_CFG_ENTRY = 'tags_page_md_path'
+_TAG_PAGE_TMPLT_CFG_ENTRY = 'tag_page_tmplt'
+_TAG_PAGE_TMPLT_PATH_CFG_ENTRY = 'tag_page_tmplt_path'
+_TAG_PAGE_MD_PATH_CFG_ENTRY = 'tag_page_md_path'
 
-_DFT_TAGS_PAGE_TMPLT = """# {{page.title}}
+_DFT_TAG_PAGE_TMPLT = """# {{page.title}}
 {% for tag in tags_and_pages %}
 ## {{tag}}
 {% for page_under_tag in tags_and_pages[tag] %}
@@ -57,9 +57,9 @@ Tags: **{{ tags | join("**, **") }}**
 
 
 class MkDocsTags(plugins.BasePlugin):
-    """A tags plugin for MkDocs
+    """An MkDocs plugin for tags support
 
-    Not intended for direct usage.
+    Not intended for direct import.
 
     Attributes:
         config_scheme: see base class
@@ -67,11 +67,11 @@ class MkDocsTags(plugins.BasePlugin):
 
     config_scheme = (
         (
-            _TAGS_PAGE_TMPLT_CFG_ENTRY,
-            config.config_options.Type(str, default=_DFT_TAGS_PAGE_TMPLT),
+            _TAG_PAGE_TMPLT_CFG_ENTRY,
+            config.config_options.Type(str, default=_DFT_TAG_PAGE_TMPLT),
         ),
         (
-            _TAGS_PAGE_TMPLT_PATH_CFG_ENTRY,
+            _TAG_PAGE_TMPLT_PATH_CFG_ENTRY,
             config.config_options.Type(str, default=''),
         ),
         (
@@ -83,33 +83,33 @@ class MkDocsTags(plugins.BasePlugin):
             config.config_options.Type(str, default=''),
         ),
         (
-            _TAGS_PAGE_MD_PATH_CFG_ENTRY,
+            _TAG_PAGE_MD_PATH_CFG_ENTRY,
             config.config_options.Type(str, default='tags.md'),
         ),
     )
 
     def __init__(self) -> None:
         self._tags_and_pages: Dict[str, Dict[str, str]] = {}
-        self._tags_page_md_path = ''
-        self._tags_page_tmplt = ''
+        self._tag_page_md_path = ''
+        self._tag_page_tmplt = ''
         self._on_page_tmplt = ''
 
     def on_config(self, config: config.Config) -> config.Config:
         """Determines the templates from the config.
 
-        See base class for arguments and return value info.
+        See base class for argument and return value info.
         """
-        # Read the path of tags page markdown file
-        self._tags_page_md_path = self.config[_TAGS_PAGE_MD_PATH_CFG_ENTRY]
-        # Read the template of the tags page
-        tags_page_tmplt_path = self.config[_TAGS_PAGE_TMPLT_PATH_CFG_ENTRY]
+        # Read the path of the markdown file of the tag page
+        self._tag_page_md_path = self.config[_TAG_PAGE_MD_PATH_CFG_ENTRY]
+        # Read the template of the tag page
+        tag_page_tmplt_path = self.config[_TAG_PAGE_TMPLT_PATH_CFG_ENTRY]
         docs_dir = config['docs_dir']
-        if tags_page_tmplt_path == '':
-            self._tags_page_tmplt = self.config[_TAGS_PAGE_TMPLT_CFG_ENTRY]
+        if tag_page_tmplt_path == '':
+            self._tag_page_tmplt = self.config[_TAG_PAGE_TMPLT_CFG_ENTRY]
         else:
-            with open(path.join(docs_dir, tags_page_tmplt_path)) as file:
-                self._tags_page_tmplt = file.read()
-        # Read the template of on-page tags lists
+            with open(path.join(docs_dir, tag_page_tmplt_path)) as file:
+                self._tag_page_tmplt = file.read()
+        # Read the template of on-page tag lists
         on_page_tmplt_path = self.config[_ON_PAGE_TMPLT_PATH_CFG_ENTRY]
         if on_page_tmplt_path == '':
             self._on_page_tmplt = self.config[_ON_PAGE_TMPLT_CFG_ENTRY]
@@ -124,9 +124,9 @@ class MkDocsTags(plugins.BasePlugin):
         config: config.Config,
         files: files.Files,
     ) -> nav.Navigation:
-        """Reads tags and title info into `self.tags_and_pages`.
+        """Reads tag and title info into `self.tags_and_pages`.
 
-        See base class for arguments and return value info.
+        See base class for argument and return value info.
         """
         config_copy = copy.copy(config)
         config_copy['plugins'] = plugins.PluginCollection()
@@ -146,7 +146,7 @@ class MkDocsTags(plugins.BasePlugin):
                             self._tags_and_pages[tag] = []
                         page_path = path.relpath(
                             page_copy.file.src_path,
-                            path.dirname(self._tags_page_md_path))
+                            path.dirname(self._tag_page_md_path))
                         self._tags_and_pages[tag].append({
                             'title': page_copy.title,
                             'path': page_path
@@ -160,25 +160,25 @@ class MkDocsTags(plugins.BasePlugin):
         config: config.Config,
         files: files.Files,
     ) -> str:
-        """Generates on-page tags lists and the tags page.
+        """Generates on-page tag lists and the tag page.
 
-        See base class for arguments and return value info.
+        See base class for argument and return value info.
         """
-        # Generate the tags page
-        if page.file.src_path == self._tags_page_md_path:
-            jinja_tmplt = jinja2.Template(self._tags_page_tmplt)
+        # Generate the tag page
+        if page.file.src_path == self._tag_page_md_path:
+            jinja_tmplt = jinja2.Template(self._tag_page_tmplt)
             markdown = jinja_tmplt.render(
                 tags_and_pages=self._tags_and_pages,
                 markdown=markdown,
                 page=page,
                 config=config,
             )
-        # Generate on-page tags lists
+        # Generate on-page tag lists
         if _TAGS_META_ENTRY not in page.meta:
             return markdown
         tags = page.meta[_TAGS_META_ENTRY]
         if isinstance(tags, list):
-            str_tags = []  # tags that are `str`
+            str_tags = []  # tags that are a `str`
             for tag in tags:
                 if isinstance(tag, str):
                     str_tags.append(tag)